Begin to rewrite code using classes
This creates two clasess, `Tradier` and `Account`. Tradier instantiates
an object with the account number and authorization token. Define these
variables in a dotenv file. Name the former variable `account_number`
and the latter variable `tradier_token`.

The second class is `Account`. It is a child class of the `Tradier`
class. It fetches info from the `Account` endpoint. Generally, it will
return an "easy to use" pandas DataFrame. It does not current include
order related endpoints, although the order endpoints are technically
in the `/account/` endpoint subdirectory.