Create Quotes class to fetch equities data
`Quotes` is a child class of the `Tradier` class. It utilizes the account
number and authorization token initialized in `Tradier`.

Endpoint URLs are defined in the `Quotes.__init__` function.

It currently has two functions, which are named `get_historical_quotes`
and `get_quote_day`. The former returns historical data about an equity
from within a date range in a pandas dataframe. The latter returns a
DataFrame row with current data about a symbol.

Still need to add a function to the `Quotes` class that will process
tick data from the gain/loss endpoint.