Add the EquityOrder class
The EquityOrder class contains functionality for placing the three main
types of orders - market, limit, and stop. It also includes a function
for stop limit orders.

There is currently an issue with placing orders when the side is set to
'short_sell'. We will need to resolve this in the future.