Create the uvatradier package distribution files
This is the first legitimate step towards organizing the tradier API
wrapper code into a structure that can be published on the PyPI repo.

All code/functionality is in uvatradier/uvatradier/tradier.py. In the
commits to follow, we will refactor tradier.py and create separate files
for each class (e.g. Tradier, Account, Quotes, etc...)