Create first functional version of uvatradier
This restructures the repository files. It organizes them to use a more
standard package-based filesystem. This is the root directory of the
`uvatradier` package uploaded to PyPI.

Previously, the repository was just a folder containing ad-hoc wrapper
functions for the tradier api. It came to be like this because we just
made and tested the files one-at-a-time. As the breadth of the code's
functionality increased, the package organization structure provided a
more sensible filesystem.