Upgrade uvatradier from 0.0.5 -> 0.0.7
Version 0.0.6 of the package was created+distributed to PyPI, but then
I encountered a merge conflict because I made a slight adjustment to
the README.md file. So, this goes straight from 5 to 7.

This commit fixes a bug on quotes.py related to importing datetime and
timedelta for the Quotes.get_historical_quotes method.